+++ js/function.js
@@ -38,3 +38,23 @@ exctractNumber('ECMAScript 2022');
 exctractNumber('1 кефир, 0.5 батона');
 exctractNumber('агент 007');
 exctractNumber('а я томат');
+
+const getRandomInt = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInt(0, elements.length - 1)];
+
+function idGenerator () {
+  let lastGeneratedId = 0;
+
+  return () => {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+}
+
+export { getRandomInt, getRandomArrayElement, idGenerator };
